Updated 11/08/2021

+++ Number Theory/Count Factors.cpp
@@ -1,6 +1,6 @@
-int countfactors(ll n) {
+int countfactors(int n) {
     int ans = 0;
-    for (ll i = 1; i*i <= n; ++i) {
+    for (int i = 1; i*i <= n; ++i) {
         if (n % i == 0) ans += 2;
         if (i*i == n) --ans;
     }
